undoSpreadsheet.cpp
The basic idea for how to collect the undo. The thing to focus here is that before the string is sent back to the client.
The string must add the version and \n to that string BEFORE being sent.
This only works with one spreadsheet and needs to modified so that the unordered map last as long as the server/spreadsheet is up.